requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.25.0
accelerate>=0.20.0
safetensors>=0.3.0
Pillow>=9.0.0
numpy>=1.21.0
pydantic>=2.0.0
python-multipart>=0.0.6
pyyaml>=6.0